+++ src/ZfSnapEventDebugger/Collector.php
@@ -14,15 +14,9 @@ class Collector extends AbstractCollector
 {
     public function collect(MvcEvent $mvcEvent)
     {
-        $sm = $mvcEvent->getApplication()->getServiceManager();
-        /* @var $moduleManager \Zend\ModuleManager\ModuleManager */
-        $moduleManager = $sm->get('ModuleManager');
+        $listener = $mvcEvent->getParam(TriggerEventListener::SELF_PARAM_NAME);
 
-        $modules = $moduleManager->getLoadedModules(false);
-
-        if (isset($modules[__NAMESPACE__]) && $modules[__NAMESPACE__] instanceof Module) {
-            $this->data = $modules[__NAMESPACE__]->getListener()->getEvents();
-        }
+        $this->data = $listener->getEvents();
     }
 
     public function getName()
